ref(server/Application/Cart/Commands): update delete request, now it's 204 status so response no needed

+++ server/CoolPizza.Application/Cart/Commands/DeleteGoodsItem.cs
@@ -5,7 +5,7 @@ using MediatR;
 
 namespace CoolPizza.Application.Cart.Commands;
 
-public class DeleteGoodsItemCommand : IRequest<DeleteCartItemDto>
+public class DeleteGoodsItemCommand : IRequest
 {
     public Guid CartId { get; init; }
     public Guid CartGoodsId { get; init; }
@@ -15,34 +15,33 @@ public class DeleteGoodsItemCommandHandler(
     IUnitOfWork unitOfWork,
     IOrdersRepository ordersRepository,
     IOrderedGoodsRepository orderedGoodsRepository
-) : IRequestHandler<DeleteGoodsItemCommand, DeleteCartItemDto>
+) : IRequestHandler<DeleteGoodsItemCommand>
 {
-    public async Task<DeleteCartItemDto> Handle(DeleteGoodsItemCommand request, CancellationToken cancellationToken)
+    public async Task Handle(DeleteGoodsItemCommand request, CancellationToken cancellationToken)
     {
         await unitOfWork.BeginTransactionAsync();
         
         try
         {
             // находим продукт корзины 
-            var cartGoodsItem = await orderedGoodsRepository.FindAsync(request.CartId, request.CartGoodsId);
+            var cartGoodsItem = await orderedGoodsRepository.FindByIdAsync(request.CartGoodsId);
             
             if (cartGoodsItem is null)
                 throw new NotFoundException("GoodsItem", request.CartGoodsId);
+            
+            // проверяем относится ли продукт к данной корзине
+            if (cartGoodsItem.OrderId != request.CartId)
+                throw new NotFoundException("CartGoods not found in this cart");
                 
             // обновляем количество
             if (!await orderedGoodsRepository.RemoveAsync(request.CartGoodsId))
                 throw new Exception("Something went wrong while deleting the cart goods.");
             
             // пересчитываем общую стоимость
-            var newTotalAmount = await ordersRepository.UpdateTotalAmount(cartGoodsItem.OrderId);
+            await ordersRepository.UpdateTotalAmount(cartGoodsItem.OrderId);
             
             // подтверждаем транзакцию
             await unitOfWork.CommitAsync();
-
-            return new DeleteCartItemDto(
-                cartGoodsItem.Id,
-                newTotalAmount
-            );
         }
         catch
         {
